requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
groq==1.1.2
pydantic==2.7.1
python-dotenv==1.0.1